+++ packages/downloading-helpers/src/file-downloads/__tests__/test-runs.spec.ts
@@ -10,6 +10,7 @@ const TEST_RUN_DATA_FILE_TYPES = [
   "coverageReplaysByFileUnmapped",
   "coverageReplaysByFileUnmappedWithRanges",
   "coverageScreenshotReplaysByFile",
+  "coverageScreenshotReplaysByFilePr",
   "coverageScreenshotReplaysByFileUnmapped",
   "coverageByReplayPr",
   "diversityByReplay",