+++ packages/remote-replay-launcher/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @alwaysmeticulous/remote-replay-launcher
 
+## 2.312.0
+
+### Patch Changes
+
+- Updated dependencies [[`d1ba630`](https://github.com/alwaysmeticulous/meticulous/commit/d1ba63009cdcb1227f9bdfe03af27a87ca7f819b), [`a45a77f`](https://github.com/alwaysmeticulous/meticulous/commit/a45a77f8157baa074cea216cdb9c620066750187), [`2b3c422`](https://github.com/alwaysmeticulous/meticulous/commit/2b3c422c47804ec7adfa79b4375c6bda7887c73c)]:
+  - @alwaysmeticulous/client@2.312.0
+  - @alwaysmeticulous/api@2.312.0
+  - @alwaysmeticulous/common@2.310.0
+
 ## 2.311.0
 
 ### Patch Changes