+++ packages/replay-orchestrator-launcher/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @alwaysmeticulous/replay-orchestrator-launcher
 
+## 2.312.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @alwaysmeticulous/downloading-helpers@2.312.0
+  - @alwaysmeticulous/common@2.310.0
+  - @alwaysmeticulous/sdk-bundles-api@2.312.0
+
 ## 2.311.0
 
 ### Patch Changes